refactor config mods tests
